+++ src/graphics/texture.rs
@@ -4,6 +4,8 @@ use std::cell::Cell;
 use std::path::Path;
 use std::rc::Rc;
 
+use image::{EncodableLayout, RgbaImage, SubImage};
+
 use crate::error::{Result, TetraError};
 use crate::fs;
 use crate::graphics::{self, DrawParams, Rectangle};
@@ -77,15 +79,8 @@ impl Texture {
     where
         P: AsRef<Path>,
     {
-        let image = fs::read_to_image(path)?.into_rgba8();
-        let (width, height) = image.dimensions();
-
-        Texture::from_rgba(
-            ctx,
-            width as i32,
-            height as i32,
-            image.into_raw().as_slice(),
-        )
+        let data = ImageData::from_file(path)?;
+        Texture::from_image_data(ctx, &data)
     }
 
     /// Creates a new texture from a slice of data, encoded in one of Tetra's supported
@@ -104,18 +99,17 @@ impl Texture {
     /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
     /// * [`TetraError::InvalidTexture`] will be returned if the texture data was invalid.
     pub fn from_file_data(ctx: &mut Context, data: &[u8]) -> Result<Texture> {
-        let image = image::load_from_memory(data)
-            .map_err(TetraError::InvalidTexture)?
-            .into_rgba8();
+        let data = ImageData::from_file_data(data)?;
+        Texture::from_image_data(ctx, &data)
+    }
 
-        let (width, height) = image.dimensions();
-
-        Texture::from_rgba(
-            ctx,
-            width as i32,
-            height as i32,
-            image.into_raw().as_slice(),
-        )
+    /// Creates a new texture from an [`ImageData`].
+    ///
+    /// # Errors
+    ///
+    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
+    pub fn from_image_data(ctx: &mut Context, data: &ImageData) -> Result<Texture> {
+        Texture::from_rgba(ctx, data.width(), data.height(), data.as_bytes())
     }
 
     /// Creates a new texture from a slice of RGBA pixel data.
@@ -128,6 +122,7 @@ impl Texture {
     ///
     /// # Errors
     ///
+    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
     /// * [`TetraError::NotEnoughData`] will be returned if not enough data is provided to fill
     /// the texture. This is to prevent the graphics API from trying to read uninitialized memory.
     pub fn from_rgba(ctx: &mut Context, width: i32, height: i32, data: &[u8]) -> Result<Texture> {
@@ -429,5 +424,156 @@ impl NineSlice {
             top: border,
             bottom: border,
         }
+    }
+}
+
+/// Raw image data.
+///
+/// Currently, the image data is always internally represented as RGBA8.
+/// This may change in the future.
+///
+/// # Supported Formats
+///
+/// Various file formats are supported, and can be enabled or disabled via Cargo features:
+///
+/// | Format | Cargo feature | Enabled by default? |
+/// |-|-|-|
+/// | PNG | `texture_png` | Yes |
+/// | JPEG | `texture_jpeg` | Yes |
+/// | GIF | `texture_gif` | Yes |
+/// | BMP | `texture_bmp` | Yes |
+/// | TIFF | `texture_tiff` | No |
+/// | TGA | `texture_tga` | No |
+/// | WebP | `texture_webp` | No |
+/// | ICO | `texture_ico` | No |
+/// | PNM | `texture_pnm` | No |
+/// | DDS/DXT | `texture_dds` | No |
+///
+/// # Performance
+///
+/// Creating or cloning an `ImageData` is a relatively expensive operation. If you can, store them in your
+/// [`State`](crate::State) struct rather than recreating them each frame.
+#[derive(Debug, Clone)]
+pub struct ImageData {
+    data: RgbaImage,
+}
+
+impl ImageData {
+    /// Loads image data from the given file.
+    ///
+    /// The format will be determined based on the file extension.
+    ///
+    /// # Errors
+    ///
+    /// * [`TetraError::FailedToLoadAsset`] will be returned if the file could not be loaded.
+    /// * [`TetraError::InvalidTexture`] will be returned if the image data was invalid.
+    pub fn from_file<P>(path: P) -> Result<ImageData>
+    where
+        P: AsRef<Path>,
+    {
+        Ok(ImageData {
+            data: fs::read_to_image(path)?.into_rgba8(),
+        })
+    }
+
+    /// Decodes image data that is encoded in one of Tetra's supported
+    /// file formats (except for TGA).
+    ///
+    /// This is useful in combination with [`include_bytes`](std::include_bytes), as it
+    /// allows you to include your image data directly in the binary.
+    ///
+    /// The format will be determined based on the 'magic bytes' at the beginning of the
+    /// data. Note that TGA files do not have recognizable magic bytes, so this function
+    /// will not recognize them.
+    ///
+    /// # Errors
+    ///
+    /// * [`TetraError::InvalidTexture`] will be returned if the image data was invalid.
+    pub fn from_file_data(data: &[u8]) -> Result<ImageData> {
+        let image = image::load_from_memory(data)
+            .map_err(TetraError::InvalidTexture)?
+            .into_rgba8();
+
+        Ok(ImageData { data: image })
+    }
+
+    /// Creates an `ImageData` from raw RGBA data.
+    ///
+    /// This function takes `Into<Vec<u8>>`. If you pass a `Vec<u8>`, that `Vec` will
+    /// be reused for the created `ImageData` without reallocating. Otherwise, the data
+    /// will be copied.
+    ///
+    /// This function requires you to provide enough data to fill the image's bounds.
+    /// If you provide too little data, an error will be returned.
+    /// If you provide too much data, it will be truncated.
+    ///
+    /// # Errors
+    ///
+    /// * [`TetraError::NotEnoughData`] will be returned if not enough data is provided to fill
+    /// the image.
+    pub fn from_rgba<D>(width: i32, height: i32, data: D) -> Result<ImageData>
+    where
+        D: Into<Vec<u8>>,
+    {
+        let data = data.into();
+
+        let expected = (width * height * 4) as usize;
+        let actual = data.len();
+
+        if actual < expected {
+            return Err(TetraError::NotEnoughData { expected, actual });
+        }
+
+        let image = RgbaImage::from_vec(width as u32, height as u32, data).unwrap();
+
+        Ok(ImageData { data: image })
+    }
+
+    /// Returns the width of the image.
+    pub fn width(&self) -> i32 {
+        self.data.width() as i32
+    }
+
+    /// Returns the height of the image.
+    pub fn height(&self) -> i32 {
+        self.data.height() as i32
+    }
+
+    /// Returns the size of the image.
+    pub fn size(&self) -> (i32, i32) {
+        let (width, height) = self.data.dimensions();
+        (width as i32, height as i32)
+    }
+
+    /// Returns the image's data, as a slice of raw bytes.
+    pub fn as_bytes(&self) -> &[u8] {
+        self.data.as_bytes()
+    }
+
+    /// Creates a new `ImageData` from a region.
+    ///
+    /// This will copy the data into a new buffer - as such, calling this function
+    /// can be expensive!
+    pub fn region(&self, region: Rectangle<i32>) -> ImageData {
+        let subimage = SubImage::new(
+            &self.data,
+            region.x as u32,
+            region.y as u32,
+            region.width as u32,
+            region.height as u32,
+        );
+
+        let data = subimage.to_image();
+
+        ImageData { data }
+    }
+
+    /// Creates a new [`Texture`] from the stored data.
+    ///
+    /// # Errors
+    ///
+    /// * [`TetraError::PlatformError`] will be returned if the underlying graphics API encounters an error.
+    pub fn to_texture(&self, ctx: &mut Context) -> Result<Texture> {
+        Texture::from_image_data(ctx, self)
     }
 }